backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Session==0.5.0
torch==2.4.1
torchvision==0.19.1
Pillow==10.0.0
Werkzeug==2.3.7